Allow explicit use of scheme-default ports in proxy URLs
If a proxy URL contains an explicit port spec ":80" (http)
or ":443" (https), Url::parse() will set the port to None, which
is indistinguishable from the case where the port is unspecified,
and will cause the port to be set to 8080.

To avoid this, if the scheme is "http" or "https", it's changed to
"xttp" or "xttps" before parsing the URL, then changed back to
the original. The bogus scheme has no known default port value, so
every explicit value will be kept.