Work around set_scheme() breakage in url-2.1.1
Since url-2.1.1, Url::set_scheme() is stricter and doesn't let us
change the scheme back to http/https. Work around the new behavior.
The changed logic also works with url < 2.1.1.

One difference from the previous version is that the string
representation of the proxy URL which was given with the explicit
scheme-default port in the environment (80 for http, 443 for https)
will no longer contain the port number.